# Add pool of hashes for HMAC