This is an experimental branch where we try to improve the performance of `hsAlg` hashers.

Current results:

```
name                 old time/op    new time/op    delta
HS/Sign-HS256-4        3.02µs ± 1%    2.80µs ± 5%   -7.30%  (p=0.000 n=10+10)
HS/Verify-HS256-4      1.53µs ± 2%    1.18µs ± 1%  -22.91%  (p=0.000 n=10+9)
HS/Sign-HS384-4        3.44µs ± 4%    2.81µs ± 4%  -18.39%  (p=0.000 n=10+9)
HS/Verify-HS384-4      1.81µs ± 1%    1.29µs ± 2%  -28.81%  (p=0.000 n=9+9)
HS/Sign-HS512-4        3.39µs ± 1%    2.87µs ± 4%  -15.24%  (p=0.000 n=9+10)
HS/Verify-HS512-4      1.86µs ± 5%    1.30µs ± 1%  -29.97%  (p=0.000 n=9+9)

name                 old alloc/op   new alloc/op   delta
HS/Sign-HS256-4        1.33kB ± 0%    0.85kB ± 0%  -36.14%  (p=0.000 n=10+10)
HS/Verify-HS256-4        512B ± 0%       32B ± 0%  -93.75%  (p=0.000 n=10+10)
HS/Sign-HS384-4        1.70kB ± 0%    0.90kB ± 0%  -47.17%  (p=0.000 n=10+10)
HS/Verify-HS384-4        848B ± 0%       48B ± 0%  -94.34%  (p=0.000 n=10+10)
HS/Sign-HS512-4        1.78kB ± 0%    0.98kB ± 0%  -45.05%  (p=0.000 n=10+10)
HS/Verify-HS512-4        864B ± 0%       64B ± 0%  -92.59%  (p=0.000 n=10+10)

name                 old allocs/op  new allocs/op  delta
HS/Sign-HS256-4          18.0 ± 0%      13.0 ± 0%  -27.78%  (p=0.000 n=10+10)
HS/Verify-HS256-4        6.00 ± 0%      1.00 ± 0%  -83.33%  (p=0.000 n=10+10)
HS/Sign-HS384-4          18.0 ± 0%      13.0 ± 0%  -27.78%  (p=0.000 n=10+10)
HS/Verify-HS384-4        6.00 ± 0%      1.00 ± 0%  -83.33%  (p=0.000 n=10+10)
HS/Sign-HS512-4          18.0 ± 0%      13.0 ± 0%  -27.78%  (p=0.000 n=10+10)
HS/Verify-HS512-4        6.00 ± 0%      1.00 ± 0%  -83.33%  (p=0.000 n=10+10)
```